Updated trident_path() to be OS Independent
Previous implementation used to get the dir name of __file__, split that using '/', exclude last element, append 'trident' to this list and join all elements. The split operation using '/' is platform dependent and breaks on Windows. One Fix is to use the current approach but substitute '/' with os.path.sep but the whole operation can be streamlined to one command.